requirements.txt
streamlit
google-generativeai
python-dotenv
langchain-community
langchain
pyPDF2
chromadb
faiss-cpu
langchain-openai
